chore: update style on menu
