+++ _mantine.scss
@@ -8,7 +8,7 @@ $mantine-breakpoint-md: "62em";
 $mantine-breakpoint-lg: "75em";
 $mantine-breakpoint-xl: "88em";
 
-@function rem($value) {
+@function toRem($value) {
   @return #{math.div(math.div($value, $value * 0 + 1), 16)}rem;
 }
