+++ src/components/footers/footer-links/footer-links.module.scss
@@ -1,17 +1,17 @@
 .footer {
-  margin-top: rem(120px);
+  margin-top: toRem(120px);
   padding-top: calc(var(--mantine-spacing-xl) * 2);
   padding-bottom: calc(var(--mantine-spacing-xl) * 2);
   background-color: light-dark(
     var(--mantine-color-gray-0),
     var(--mantine-color-dark-6)
   );
-  border-top: rem(1px) solid
+  border-top: toRem(1px) solid
     light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-5));
 }
 
 .logo {
-  max-width: rem(200px);
+  max-width: toRem(200px);
 
   @media (max-width: $mantine-breakpoint-sm) {
     display: flex;
@@ -21,7 +21,7 @@
 }
 
 .description {
-  margin-top: rem(5px);
+  margin-top: toRem(5px);
 
   @media (max-width: $mantine-breakpoint-sm) {
     margin-top: var(--mantine-spacing-xs);
@@ -49,15 +49,15 @@
 }
 
 .wrapper {
-  width: rem(160px);
+  width: toRem(160px);
 }
 
 .link {
   display: block;
   color: light-dark(var(--mantine-color-gray-6), var(--mantine-color-dark-1));
   font-size: var(--mantine-font-size-sm);
-  padding-top: rem(3px);
-  padding-bottom: rem(3px);
+  padding-top: toRem(3px);
+  padding-bottom: toRem(3px);
 
   &:hover {
     text-decoration: underline;
@@ -81,7 +81,7 @@
   margin-top: var(--mantine-spacing-xl);
   padding-top: var(--mantine-spacing-xl);
   padding-bottom: var(--mantine-spacing-xl);
-  border-top: rem(1px) solid
+  border-top: toRem(1px) solid
     light-dark(var(--mantine-color-gray-2), var(--mantine-color-dark-4));
 
   @media (max-width: $mantine-breakpoint-sm) {
