+++ src/components/headers/header-mega-menu/header-mega-menu.module.scss
@@ -1,8 +1,8 @@
 .header {
-  height: rem(60px);
+  height: toRem(60px);
   padding-left: var(--mantine-spacing-md);
   padding-right: var(--mantine-spacing-md);
-  border-bottom: rem(1px) solid
+  border-bottom: toRem(1px) solid
     light-dark(var(--mantine-color-gray-3), var(--mantine-color-dark-4));
 }
 
@@ -18,9 +18,8 @@
   font-size: var(--mantine-font-size-sm);
 
   @media (max-width: $mantine-breakpoint-sm) {
-    height: rem(42px);
+    height: toRem(42px);
     width: 100%;
-    background-color: red;
   }
 
   @include hover {
@@ -53,6 +52,6 @@
   margin-top: var(--mantine-spacing-sm);
   padding: var(--mantine-spacing-md) calc(var(--mantine-spacing-md) * 2);
   padding-bottom: var(--mantine-spacing-xl);
-  border-top: rem(1px) solid
+  border-top: toRem(1px) solid
     light-dark(var(--mantine-color-gray-1), var(--mantine-color-dark-5));
 }
