+++ src/components/headers/header-mega-menu/header-mega-menu.tsx
@@ -20,49 +20,49 @@ export function HeaderMegaMenu({
   menuItems,
   subMenuContent,
   rightContent,
+  logo,
 }: HeaderWithMegaMenuProps) {
   const [drawerOpened, { toggle: toggleDrawer, close: closeDrawer }] =
     useDisclosure(false)
-  const [, { toggle: toggleLinks }] = useDisclosure(false)
   const theme = useMantineTheme()
 
   return (
     <Box pb={120}>
       <header className={classes.header}>
         <Group justify="space-between" h="100%">
+          {logo}
           <Group h="100%" gap={0} visibleFrom="sm">
-            {menuItems.map((item) => {
+            {menuItems.map((item, index) => {
               return (
-                <>
-                  <HoverCard
-                    width={600}
-                    position="bottom"
-                    radius="md"
-                    shadow="md"
-                    withinPortal
-                  >
-                    <HoverCard.Target>
-                      <a href="#" className={classes.link}>
-                        <Center inline>
-                          <Box component="span" mr={5}>
-                            {item.title}
-                          </Box>
+                <HoverCard
+                  width={600}
+                  position="bottom"
+                  radius="md"
+                  shadow="md"
+                  withinPortal
+                  key={index}
+                >
+                  <HoverCard.Target>
+                    <a href={item.link} className={classes.link}>
+                      <Center inline>
+                        <Box component="span" mr={5}>
+                          {item.title}
+                        </Box>
+                        {subMenuContent && (
                           <IconChevronDown
                             style={{ width: rem(16), height: rem(16) }}
                             color={theme.colors.blue[6]}
                           />
-                        </Center>
-                      </a>
-                    </HoverCard.Target>
-                    {subMenuContent}
-                  </HoverCard>
-                </>
+                        )}
+                      </Center>
+                    </a>
+                  </HoverCard.Target>
+                  {subMenuContent}
+                </HoverCard>
               )
             })}
           </Group>
-
-          {rightContent}
-
+          {rightContent && <Group visibleFrom="sm">{rightContent}</Group>}
           <Burger
             opened={drawerOpened}
             onClick={toggleDrawer}
@@ -75,17 +75,17 @@ export function HeaderMegaMenu({
         opened={drawerOpened}
         onClose={closeDrawer}
         size="100%"
-        padding="md"
+        padding="lg"
         title="Navigation"
         hiddenFrom="sm"
         zIndex={1_000_000}
       >
-        <ScrollArea h={`calc(100vh - ${rem(80)})`} mx="-md">
+        <ScrollArea h={`calc(100vh - ${rem(80)})`}>
           <Divider my="sm" />
-          {menuItems.map((item) => {
+          {menuItems.map((item, index) => {
             return (
-              <>
-                <UnstyledButton className={classes.link} onClick={toggleLinks}>
+              <div key={index}>
+                <UnstyledButton className={classes.link}>
                   <Center inline>
                     <Box component="span" mr={5}>
                       {item.title}
@@ -99,13 +99,15 @@ export function HeaderMegaMenu({
                   </Center>
                 </UnstyledButton>
                 {subMenuContent}
-              </>
+              </div>
             )
           })}
-
           <Divider my="sm" />
-
-          {rightContent}
+          {rightContent && (
+            <Group justify="center" grow pb="xl" px="md">
+              {rightContent}
+            </Group>
+          )}
         </ScrollArea>
       </Drawer>
     </Box>
